test: 사용자가 가진 voucher 찾기 테스트 및 @nested 방식 도입
